requirements.txt
altgraph==0.17.4
build==0.9.0
certifi==2024.2.2
charset-normalizer==2.0.12
datacryptchain==0.0.0.9008
editables==0.2
hatchling==0.22.0
idna==3.6
importlib-metadata==4.8.3
Naked==0.1.32
packaging==21.3
pathspec==0.9.0
pep517==0.13.1
pkg_resources==0.0.0
pluggy==1.0.0
pyasn1==0.5.1
pycryptodomex==3.20.0
pyfakefs==4.6.3
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.27.1
rsa==4.9
shellescape==3.8.1
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.18
zipp==3.6.0
